Correct the reference to secretsmanager secret
Serverless only allows access to AWS secretsmanager secret via SSM.
However this imposes restrictions on the name of the secretsmanager
secret as well as it's format.

A secretsmanager secret named to be accessible to concourse pipelines
usually is of the form /concourse/$team-name/$secret-name (for
team-wide secrets). This secret cannot be directly used in Serverless
since to access the secret, serverless goes via Paramter Store which
cannot resolve the reference due to slashes in the name.

The fix is to create a secretsmanager that:
1. Does not have forbidden characters in it's name
2. Is not stored as Binary