Update brk_serverless.yml
The changes to make apply factors and calculate movement use unique_identifier instead of reference fell down in bricks because it is used in a different way. I've added responder_id to the unique_identifier variable in the config, this change is so that its ignored by the modules using the other unique identifiers.